1.6.4 Fait : - Lister les produit dans le webservice ok - Détail des produits Ok - Ajout dans le panier + retrait Ok
A faire :
- A l'appui sur le bouton back dans productDetailFragment l'appli plante.
- Impossible d'afficher le panier
- Ajouter Stripe
- Ajouter la map de la zone livré et bloquer le user si pas dans la zone

+++ app/src/main/java/com/apero_area/aperoarea/view/activities/MainActivity.java
@@ -124,7 +124,7 @@ public class MainActivity extends AppCompatActivity {
                     .getListOfProductsInShoppingList()) {
 
                 updateCheckOutAmount(
-                        BigDecimal.valueOf(Long.valueOf(product.getSellMRP())),
+                        BigDecimal.valueOf(Double.valueOf(product.getSellMRP())),
                         true);
             }
         }
